pyproject.toml
[project]
dependencies = [
  "arxiv>=2.2.0",
  "faiss-cpu>=1.12.0",
  "fitz>=0.0.1.dev2",
  "langchain-cohere>=0.4.6",
  "langchain-community>=0.3.30",
  "langchain-experimental>=0.3.4",
  "langchain-openai>=0.3.35",
  "langchain-tavily>=0.2.11",
  "langchain-teddynote>=0.5.1",
  "langchain-text-splitters>=0.3.11",
  "langgraph>=0.6.8",
  "langgraph-api>=0.4.43",
  "langgraph-checkpoint-sqlite>=2.0.11",
  "langgraph-cli[inmem]>=0.4.4",
  "langgraph-supervisor>=0.0.29",
  "langgraph-swarm>=0.0.14",
  "matplotlib>=3.10.7",
  "notebook>=7.4.7",
  "pymupdf>=1.26.5",
  "python-dotenv>=1.1.1",
  "streamlit>=1.50.0",
  "trustcall>=0.0.39",
  "unstructured>=0.18.15",
  "wikipedia>=1.4.0",
]
description = "Add your description here"
name = "foundation-introduction-to-langgraph"
readme = "README.md"
requires-python = ">=3.12"
version = "0.1.0"

[dependency-groups]
dev = [
  "ipykernel>=6.30.1,<7.0.0",
]
